requirements.txt
selenium==4.18.1
python-docx==1.1.0